Fixed and improved `p2mm_startsession` CON_COMMAND
Fixed the `p2mm_startsession` CON_COMMAND to properly start a explicitly requested session and not still load and map specified in the `p2mm_lastmap` ConVar.

Also updated the `plugin_print` description of the plugin to include adjustable version values and logging to take the value of `p2mm_developer` and not the normal `developer` ConVar.