Multiple changes involving game events and VScript handles
Replaced tracking portal_player_touchedground game event with player_landed game event because that one actually works.

Added "bugpause" and "bugunpause" to the list of banned client commands that clients can't execute.

Removed not working VScript call for OnPlayerJoin in the "player_connect" game event because its to early of a time for the plugin to be getting information about the player entity.

Added "game_info" game event.

Added ClientActive plugin callback which is where OnPlayerJoin will be called and where the last map value will be set.