Added RespawnPlayer and SetFlashlightState functions
Renamed RemoveEntity to match the naming scheme for functions in Portal 2 classes.
Renamed CBasePlayer__PlayerIndexToPlayer because it is not a class function of CBasePlayer, instead it's UTIL_PlayerByIndex.

Added CPortal_Player__RespawnPlayer and CPortal_Player__SetFlashlightState as usable functions both in the plugin and in VScript, RespawnPlayer and SetFlashlightState respectively.