add disclaimer, fixed bugs
Added a warning if the webhook variable is set so players are informed of the chat being recorded.
Added a timestamp so the Discord play timer is preserved throughout level changes.
Fixed returning if webhook is not set.
Removed "Starting Map" being sent to RPC as it is a waste of the rate limit.