Cleaned up and updated parts of Discord RPC
Instead of making a DiscordRichPresence struct in memory everytime to update the current presence, changed it so there is one RPC struct in the class which is changed then Discord_UpdatePresence is called with it. That way any information we wanted to keep the same can stay the same.

Also added lists of P2 and PS:Mel maps which will be listed through to get name information and chapter information on the map. Using this information specific smallImageKeys and smallImageText values can be set in the RPC. Some of them are working, except the chapter specific ones are not, and I am unsure why but will investigate later.