Some more game support things, supported mod checking, changing some if statements to switch statements, and fixed Discord RPC bug
Added in some more support things for Divinity like Discord RPC and a place for the largeImageKey when that is completed.

Reworked the part which originally just printed a dev message of which Portal 2 branch based game is being played to checking if it is currently a supported game. If it isn't the game will be closed because half of the time the game will close anyway because of crashing due to wrong byte patches or hooks, or some other inconvenience we haven't fully figured out. If you want to load in anyway just use `-forcep2mmload` as a launch option in whatever you use to launch the game.

Because the m_iCurGameIndex has a enum to go with it, switched out some if statement checks with switch case statements.

Fixed Discord RPC bug where it wouldnt report on discord that the user was actually on the menu when exiting out of a game session.